Moved plugin tools to tools subdirectory
